+++ types/src/lib.rs
@@ -45,7 +45,8 @@ pub use crate::share::*;
 pub use crate::sync::*;
 pub use crate::validate::*;
 
-// uniffi :(
+// `uniffi::use_remote_type` macro seems a bit, limited in that it it work correctly only
+// for types that are exported in the root of the crate
 #[cfg(feature = "uniffi")]
 pub use crate::hash::Hash;
